TransparentCharacterCreation/requirements.txt
openai
cv2
Pillow